extract modifier types to central location
